Make InjKit work for kotlin
Summary:
We want to make sure `AutoHandleExceptions` annotations also work with Kotlin to make sure migration is smooth and does not have much changes in behaviour.

`apply-plugin` is needed in order for kotlin-tasks in gradle to exist, so we can use them in `safekit-build.gradle` when we try to get `kotlinProvider`. This is because the safekit-build is in the root and applies to all plugins. _Maybe_ it is possible to refactor the existing stuff and copypaste the same config everywhere, but I do not think it is worth the time.

The addition in InjKit is similar to invoke-method without class-loading stuff. I am not sure why it is needed, it will recursively call itself after, I tried to go back in history and I suspect it couldve been some bug in gradle. Not having it works as well. I also did not want to ruthlessly refactor it since I do not know how many might use it as it is, and easier to add a new method that does almost same things and imo it is still clean.

KotlinTask does not call getBootstrapPath, reason for that is I have not been able to find a method like it when I used reflection. To repro:

```
//    System.out.println("task clazz " + compileTask.getClass());
//    Method[] declaredMethods = compileTask.getClass().getDeclaredMethods();

//    for (Method m : declaredMethods) {
//      System.out.println(m.toString());
//    }
//    System.out.println(compileTask.getClass().getDeclaredMethods());
//    System.out.println("XXXXX");
//    Object getKotlinOptions = invokeGetter(compileTask, "getKotlinOptions");
//    Method[] declaredMethods1 = getKotlinOptions.getClass().getDeclaredMethods();
//    for (Method m : declaredMethods1) {
//      System.out.println(m.toString());
//    }
```

Reviewed By: ashaevy

Differential Revision: D26286342

fbshipit-source-id: 3f7b39b2d1d073b8718f3dc5aa56b85fb2e0f179